Fix bug with seed 1354
The adaptive parameters were not working well with seed 1354 and a correct answer was being marked incorrect.  I suspect that having two variables in the context (x and u) contributed to these problems.  I rewrote the problem so that the unnecessary intermediate variable u is not in the context and also adjusted the domain of function evaluation to make the answer checker more robust.